+++ sfall/ddraw.vcxproj
@@ -345,6 +345,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="CheckAddress.h" />
+    <ClInclude Include="ConsoleWindow.h" />
     <ClInclude Include="Delegate.h" />
     <ClInclude Include="FalloutEngine\AsmMacros.h" />
     <ClInclude Include="FalloutEngine\Enums.h" />
@@ -496,6 +497,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="CheckAddress.cpp" />
+    <ClCompile Include="ConsoleWindow.cpp" />
     <ClCompile Include="FalloutEngine\EngineUtils.cpp" />
     <ClCompile Include="FalloutEngine\FunctionOffsets.cpp" />
     <ClCompile Include="FalloutEngine\Variables.cpp" />
