+++ sfall/main.h
@@ -81,10 +81,6 @@ namespace sfall
 // Trap for Debugger
 #define BREAKPOINT __asm int 3
 
-// Macros for quick replacement of assembler opcodes pushad/popad
-#define pushadc __asm push eax __asm push edx __asm push ecx
-#define popadc __asm pop ecx __asm pop edx __asm pop eax
-
 extern bool versionCHI;
 
 extern char falloutConfigName[65];
